Refactor load game dialog, settings views, and game list widget into modular components, and fix theme refresh issues
