+++ src/gui/components/__init__.py
@@ -2,5 +2,6 @@
 Shared GUI components module.
 """
 from .stat_card import StatCard, SimpleStatCard
+from .masonry_layout import MasonryLayout
 
-__all__ = ['StatCard', 'SimpleStatCard']
+__all__ = ['StatCard', 'SimpleStatCard', 'MasonryLayout']
