+++ src/gui/views/settings/__init__.py
@@ -1,0 +1,19 @@
+"""
+Settings view modules.
+"""
+from .engine_settings import EngineSettings
+from .api_settings import ApiSettings, test_llm_sync
+from .player_settings import PlayerSettings
+from .appearance_settings import AppearanceSettings
+from .data_settings import DataSettings
+from .links_settings import LinksSettings
+
+__all__ = [
+    'EngineSettings',
+    'ApiSettings',
+    'PlayerSettings',
+    'AppearanceSettings',
+    'DataSettings',
+    'LinksSettings',
+    'test_llm_sync',
+]
